+++ cqrs-core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # master
 
+* No changes yet
+
+# [[0.2.2] 2019-09-05](https://github.com/cq-rs/cqrs/releases/tag/cqrs-core-0.2.2)
+
 * Add new type `Before` for reading event streams backward
 
 # [[0.2.1] 2019-04-29](https://github.com/cq-rs/cqrs/releases/tag/cqrs-core-0.2.1)